% Auto-generated: do not edit by hand
\name{datepickerByInclusion}

\alias{datepickerByInclusion}

\title{DatepickerByInclusion component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
datepickerByInclusion(id=NULL, date=NULL, datesIncluded=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{date}{Character. The value displayed in the input.}

\item{datesIncluded}{Unnamed list. The value displayed in the input.}
}
